// src/api-client/models/expense-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Financial Aide Backend
 * API Backend for Financial Aide open-source budgeting system
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BudgetCategoryResponse } from './budget-category-response';
import { UserResponse } from './user-response';
/**
 * 
 * @export
 * @interface ExpenseResponse
 */
export interface ExpenseResponse {
    /**
     * 
     * @type {string}
     * @memberof ExpenseResponse
     */
    name?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof ExpenseResponse
     */
    timestamp: Date;
    /**
     * 
     * @type {UserResponse}
     * @memberof ExpenseResponse
     */
    user: UserResponse;
    /**
     * 
     * @type {string}
     * @memberof ExpenseResponse
     */
    description?: string | null;
    /**
     * 
     * @type {BudgetCategoryResponse}
     * @memberof ExpenseResponse
     */
    category: BudgetCategoryResponse;
    /**
     * 
     * @type {string}
     * @memberof ExpenseResponse
     */
    amount: string;
    /**
     * 
     * @type {number}
     * @memberof ExpenseResponse
     */
    id: number;
}